test(svg): isolate #159 SVG render tests in their own spec file, run on wasi
Replaces the native-only skip (e964fb2). The #159 from_svg() render logic is
verified-correct on wasm: the actual CI-built wasm32-wasip1-threads binary
passes every SVG test in isolation. The wasi failures were never the render
code -- they are a pre-existing wasm-threads fault: the async avif()/webp()
codec tests run aom's deeply-recursive AVIF encode on a SPAWNED WORKER thread,
whose stack uses wasi-libc's small default. The -C link-arg=-zstack-size=16MiB
flag (.cargo/config.toml) sizes only the MAIN thread's shadow stack, so aom
overflows the worker stack into the shared linear memory and silently corrupts
the heap; a later SVG render in the same instance then reads the corrupted
state and rasterizes a corner speck (x86 CI) / a bogus multi-TB allocation ->
SIGABRT, or faults "memory access out of bounds" (arm64). Flaky, host-layout
dependent.

ava runs each test file in its own process (workerThreads: false), so moving
the heavy SVG render-logic regressions into transformer-svg.spec.mjs gives them
a dedicated wasm instance that never spawns the codec worker -> no corruption.
They now run unconditionally on wasi, restoring genuine wasi coverage of the
#159 fix. transformer.spec.mjs keeps the codec tests + one jpegSync SVG smoke
test. Verified on arm64: 20/20 SVG-only clean and 15/15 both-files clean (vs
~10% crash when unified). The worker-stack overflow itself is a pre-existing
wasm-threads limitation, tracked separately.
